refactor: organize component directory structure
